🦄 refactor: Adjunting file
Product carousel has in a empty folder
