+++ src/app/components/productCarousel/product-carousel.component.spec.ts
